api/requirements.txt
# anyio 4 라인에 맞춘 호환 세트
a2a-sdk==0.3.0
fastapi==0.115.0
sse-starlette==3.0.2

openai==1.101.0
httpx==0.28.1
pydantic==2.11.7
jinja2==3.1.6
markdown==3.8.2
python-dotenv==1.1.1
aiofiles==24.1.0
python-multipart==0.0.20
